Оптимизировала кол-во пакетов, изменила названия в соответствии с правилами нейминга.
