+++ allure-report/data/attachments/5ac4ec6671837782.txt
@@ -1,4 +1,4 @@
 -------------------------------------------------------------------------------
-Test set: sprint7.NumberOfOrders.OrdersNumberTest
+Test set: project.ordertests.OrdersNumberTest
 -------------------------------------------------------------------------------
 Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.215 sec
