+++ allure-report/data/attachments/9d482a546572947f.txt
@@ -1,4 +1,4 @@
 -------------------------------------------------------------------------------
-Test set: sprint7.CourierLogin.CourierLoginTest
+Test set: project.couriertests.CourierLoginTest
 -------------------------------------------------------------------------------
 Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.882 sec
