+++ allure-report/data/attachments/9f9869d27b680e17.txt
@@ -1,4 +1,4 @@
 -------------------------------------------------------------------------------
-Test set: sprint7.OrderCreation.CreateOrderParametrizedTest
+Test set: project.ordertests.CreateOrderParametrizedTest
 -------------------------------------------------------------------------------
 Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.316 sec
