+++ allure-report/data/attachments/afa78c31a8e071c8.txt
@@ -1,4 +1,4 @@
 -------------------------------------------------------------------------------
-Test set: sprint7.CourierCreation.CourierCreationTest
+Test set: project.couriertests.CourierCreationTest
 -------------------------------------------------------------------------------
 Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.267 sec
