+++ allure-report/data/attachments/f94a50ec09bc33eb.txt
@@ -1,4 +1,4 @@
 -------------------------------------------------------------------------------
-Test set: sprint7.OrderCreation.CreateOrderTest
+Test set: project.ordertests.CreateOrderTest
 -------------------------------------------------------------------------------
 Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.1 sec
