+++ allure-report/data/suites.csv
@@ -1,14 +1,14 @@
 "Status","Start Time","Stop Time","Duration in ms","Parent Suite","Suite","Sub Suite","Test Class","Test Method","Name","Description"
-"passed","","","216","","sprint7.CourierCreation.CourierCreationTest","","sprint7.CourierCreation.CourierCreationTest","","creationFailsWithTheSameLogin",""
-"passed","","","417","","sprint7.CourierLogin.CourierLoginTest","","sprint7.CourierLogin.CourierLoginTest","","loggedInReturnId",""
-"passed","","","216","","sprint7.CourierLogin.CourierLoginTest","","sprint7.CourierLogin.CourierLoginTest","","loginFailsWithoutLogin",""
-"passed","","","133","","sprint7.CourierLogin.CourierLoginTest","","sprint7.CourierLogin.CourierLoginTest","","fakeCredsLoggInFailed",""
-"passed","","","100","","sprint7.OrderCreation.CreateOrderTest","","sprint7.OrderCreation.CreateOrderTest","","orderCreation",""
-"passed","","","133","","sprint7.CourierCreation.CourierCreationTest","","sprint7.CourierCreation.CourierCreationTest","","courierCreatedSuccesfully",""
-"passed","","","2718","","sprint7.CourierCreation.CourierCreationTest","","sprint7.CourierCreation.CourierCreationTest","","creationFailsWithoutPassword",""
-"passed","","","3215","","sprint7.NumberOfOrders.OrdersNumberTest","","sprint7.NumberOfOrders.OrdersNumberTest","","getOrderNumberByTrack",""
-"passed","","","100","","sprint7.OrderCreation.CreateOrderParametrizedTest","","sprint7.OrderCreation.CreateOrderParametrizedTest","","orderCreationWithColors[2]",""
-"passed","","","117","","sprint7.OrderCreation.CreateOrderParametrizedTest","","sprint7.OrderCreation.CreateOrderParametrizedTest","","orderCreationWithColors[1]",""
-"passed","","","99","","sprint7.OrderCreation.CreateOrderParametrizedTest","","sprint7.OrderCreation.CreateOrderParametrizedTest","","orderCreationWithColors[0]",""
-"passed","","","100","","sprint7.CourierLogin.CourierLoginTest","","sprint7.CourierLogin.CourierLoginTest","","loginFailsWithoutPassword",""
-"passed","","","133","","sprint7.CourierCreation.CourierCreationTest","","sprint7.CourierCreation.CourierCreationTest","","creationFailsWithoutLogin",""
+"passed","","","216","","project.couriertests.CourierCreationTest","","project.couriertests.CourierCreationTest","","creationFailsWithTheSameLogin",""
+"passed","","","417","","project.couriertests.CourierLoginTest","","project.couriertests.CourierLoginTest","","loggedInReturnId",""
+"passed","","","216","","project.couriertests.CourierLoginTest","","project.couriertests.CourierLoginTest","","loginFailsWithoutLogin",""
+"passed","","","133","","project.couriertests.CourierLoginTest","","project.couriertests.CourierLoginTest","","fakeCredsLoggInFailed",""
+"passed","","","100","","project.ordertests.CreateOrderTest","","project.ordertests.CreateOrderTest","","orderCreation",""
+"passed","","","133","","project.couriertests.CourierCreationTest","","project.couriertests.CourierCreationTest","","courierCreatedSuccesfully",""
+"passed","","","2718","","project.couriertests.CourierCreationTest","","project.couriertests.CourierCreationTest","","creationFailsWithoutPassword",""
+"passed","","","3215","","project.ordertests.OrdersNumberTest","","project.ordertests.OrdersNumberTest","","getOrderNumberByTrack",""
+"passed","","","100","","project.ordertests.CreateOrderParametrizedTest","","project.ordertests.CreateOrderParametrizedTest","","orderCreationWithColors[2]",""
+"passed","","","117","","project.ordertests.CreateOrderParametrizedTest","","project.ordertests.CreateOrderParametrizedTest","","orderCreationWithColors[1]",""
+"passed","","","99","","project.ordertests.CreateOrderParametrizedTest","","project.ordertests.CreateOrderParametrizedTest","","orderCreationWithColors[0]",""
+"passed","","","100","","project.couriertests.CourierLoginTest","","project.couriertests.CourierLoginTest","","loginFailsWithoutPassword",""
+"passed","","","133","","project.couriertests.CourierCreationTest","","project.couriertests.CourierCreationTest","","creationFailsWithoutLogin",""
