+++ src/test/java/project/helpers/Courier.java
@@ -1,20 +1,17 @@
-package sprint7.Helpers;
+package project.helpers;
 
-public class Credentials {
+public class Courier {
     private String login;
-    // ключ link стал полем типа String
+
     private String password;
+    private String firstName;
 
 
-    // конструктор со всеми параметрами
-    public Credentials(String login, String password) {
+    public Courier(String login, String password, String firstName) {
         this.login = login;
         this.password = password;
+        this.firstName = firstName;
     }
-    public static Credentials from(Courier courier){
-        return new Credentials(courier.getLogin(), courier.getPassword());
-    }
-
 
     public String getLogin() {
         return login;
@@ -24,7 +21,9 @@ public class Credentials {
         return password;
     }
 
-
+    public String getFirstName() {
+        return firstName;
+    }
 
     public void setLogin(String login) {
         this.login = login;
@@ -34,9 +33,11 @@ public class Credentials {
         this.password = password;
     }
 
-
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
 
     // конструктор без параметров
-    public Credentials() {
+    public Courier() {
     }
 }
