+++ src/test/java/project/helpers/OrderNoColor.java
@@ -1,8 +1,6 @@
-package sprint7.Helpers;
+package project.helpers;
 
-import java.util.List;
-
-public class Order {
+public class OrderNoColor {
     private String firstName;
 
     private String lastName;
@@ -13,10 +11,12 @@ public class Order {
     private String rentTime;
     private String deliveryDate;
     private String comment;
+    private String color;
 
 
 
-    public Order(String firstName, String lastName, String address, String metroStation, String phone, String rentTime, String deliveryDate, String comment, List<String> colors) {
+
+    public OrderNoColor(String firstName, String lastName, String address, String metroStation, String phone, String rentTime, String deliveryDate, String comment, String color) {
 
         this.firstName = firstName;
         this.lastName = lastName;
@@ -26,6 +26,7 @@ public class Order {
         this.rentTime = rentTime;
         this.deliveryDate = deliveryDate;
         this.comment = comment;
+        this.color = color;
 
 
 
@@ -95,9 +96,15 @@ public class Order {
         this.comment = comment;
     }
 
+    public String getColor() {
+        return color;
+    }
 
+    public void setColor(String color) {
+        this.color = color;
+    }
 
     // конструктор без параметров
-    public Order() {
+    public OrderNoColor() {
     }
 }
